profile_manager/requirements.txt
django >=1.9.1, <2.0
Pillow
cognitive_face